+++ cookbooks/rabbitmq/metadata.rb
@@ -1,15 +1,17 @@
+name              "rabbitmq"
 maintainer        "Opscode, Inc."
 maintainer_email  "cookbooks@opscode.com"
 license           "Apache 2.0"
 description       "Installs and configures RabbitMQ server"
-version           "1.6.3"
+version           "2.1.2"
 recipe            "rabbitmq", "Install and configure RabbitMQ"
 recipe            "rabbitmq::cluster", "Set up RabbitMQ clustering."
-depends           "apt", ">= 1.4.4"
-depends           "yum", ">= 0.5.0"
+recipe            "rabbitmq::plugin_management", "Manage plugins with node attributes"
+recipe            "rabbitmq::virtualhost_management", "Manage virtualhost with node attributes"
+recipe            "rabbitmq::user_management", "Manage users with node attributes"
 depends           "erlang", ">= 0.9"
 
-%w{ubuntu debian redhat centos scientific amazon fedora}.each do |os|
+%w{ubuntu debian linuxmint redhat centos scientific amazon fedora oracle smartos}.each do |os|
   supports os
 end
 
@@ -62,3 +64,33 @@ attribute "rabbitmq/cluster_disk_nodes",
 attribute "rabbitmq/erlang_cookie",
   :display_name => "RabbitMQ Erlang cookie",
   :description => "Access cookie for clustering nodes.  There is no default."
+
+attribute "rabbitmq/virtualhosts",
+  :display_name => "Virtualhosts on rabbitmq instance",
+  :description => "List all virtualhosts that will exist",
+  :default => [],
+  :type => "array"
+
+attribute "rabbitmq/enabled_users",
+  :display_name => "Users and their rights on rabbitmq instance",
+  :description => "Users and description of their rights",
+  :default => [{ :name => "guest", :password => "guest", :rights => [{:vhost => nil , :conf => ".*", :write => ".*", :read => ".*"}]}],
+  :type => "array"
+
+attribute "rabbitmq/disabled_users",
+  :display_name => "Disabled users",
+  :description => "List all users that will be deactivated",
+  :default => [],
+  :type => "array"
+
+attribute "rabbitmq/enabled_plugins",
+  :display_name => "Enabled plugins",
+  :description => "List all plugins that will be activated",
+  :default => [],
+  :type => "array"
+
+attribute "rabbitmq/disabled_plugins",
+  :display_name => "Disabled plugins",
+  :description => "List all plugins that will be deactivated",
+  :default => [],
+  :type => "array"