+++ cookbooks/rabbitmq/resources/user.rb
@@ -2,7 +2,7 @@
 # Cookbook Name:: rabbitmq
 # Resource:: user
 #
-# Copyright 2011, Opscode, Inc.
+# Copyright 2011-2013, Opscode, Inc.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -17,12 +17,13 @@
 # limitations under the License.
 #
 
-actions :add, :delete, :set_permissions, :clear_permissions
+actions :add, :delete, :set_permissions, :clear_permissions, :set_tags, :clear_tags
 
 attribute :user, :kind_of => String, :name_attribute => true
 attribute :password, :kind_of => String
 attribute :vhost, :kind_of => String
 attribute :permissions, :kind_of => String
+attribute :tag, :kind_of => String
 
 def initialize(*args)
   super